+++ src/amplitude-snippet.js
@@ -4,6 +4,7 @@
   as.type = 'text/javascript';
   as.async = true;
   as.src = 'https://d24n15hnbwhuhn.cloudfront.net/libs/amplitude-2.4.1-min.gz.js';
+  as.onload = function() { window.amplitude.runQueuedFunctions(); };
   var s = document.getElementsByTagName('script')[0];
   s.parentNode.insertBefore(as, s);
   var Identify = function() { this._q = []; return this; };